+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
   * new backend for Openrunner
   * gpxdo --similar now supports a list of tracks
   * new: gpxdo --join
+  * Track.merge (and gpxdo merge) now also merge waypoints
 
 1.4.1 release 2018-10-22
 ------------------------